criacao do ParseUtils
