+++ back/tasks/src/main/java/br/com/caio/todo/tasks/vo/TasksVO.java
@@ -42,27 +42,7 @@ public class TasksVO {
 		this.userName = userName;
 	}
 
-	public static List<TasksVO> parse(List<Tasks> taskList) {
-		List<TasksVO> result = new ArrayList<>();
-		for (Tasks task : taskList) {
-			TasksVO tasksVO = new TasksVO(task.getId(), 
-					task.getNameTask(), task.getDescription(), task.getStatusTask(), task.getCreationDate(), task.getCompletedDate(), 
-					task.getDeadlineDate(), task.getUser().getUserName());
-			result.add(tasksVO);
-		}
-		
-		return result;
-	}
-	
-	public static Tasks parseToEntity(TasksVO tasksVO, User user) {
-		Tasks tasks = new Tasks();
-		tasks.setNameTask(tasksVO.getNameTask());
-		tasks.setDescription(tasksVO.getDescription());
-		tasks.setDeadlineDate(tasksVO.getDeadlineDate());
-		tasks.setUser(user);
-		
-		return tasks;
-	}
+
 
 	public Integer getId() {
 		return id;
